project/requirements.txt
amqp==5.2.0
APScheduler==3.9.1
asgiref==3.8.1
asn1crypto==1.5.1
autocommand==2.2.1
beautifulsoup4==4.11.1
bots-ediint==3.3.1
cachetools==5.5.0
certifi==2020.6.20
cffi
chardet==3.0.4
charset-normalizer==2.1.1
cheroot==8.6.0
CherryPy==18.8.0
ci-info==0.3.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
clr-loader==0.1.7
configobj==5.0.8
configparser==7.0.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.1
cssutils==1.0.2
Django==5.1.4
django-background-task==0.1.8
django-cors-headers==3.11.0
django-filter==24.3
django-js-asset==2.0.0
django-mathfilters==1.0.0
djangorestframework==3.15.2
djangorestframework-api-key==3.0.0
djangorestframework-datatables==0.5.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
ecdsa==0.18.0
ellipticpy==0.0.1
etelemetry==0.3.1
eth-hash==0.3.3
eth-keys==0.4.0
eth-typing==3.2.0
eth-utils==2.0.0
filelock==3.14.0
fitz==0.0.1.dev2
google-api-core==2.20.0
google-api-python-client==2.147.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.65.0
httplib2==0.22.0
idna==3.3
incremental==17.5.0
inflect==6.0.0
inflection==0.5.1
isodate==0.6.1
itypes==1.2.0
jaraco.classes==3.2.2
jaraco.collections==3.5.2
jaraco.context==4.1.2
jaraco.functools==3.5.1
jaraco.text==3.9.1
Jinja2==3.1.4
josepy==1.14.0
JPype1==1.5.0
kombu==5.3.5
looseversion==1.3.0
lxml
MarkupSafe==2.1.5
more-itertools==8.14.0
mozilla-django-oidc==4.0.1
msgpack==1.0.0
networkx==3.3
nibabel==5.2.1
nipype==1.8.6
numpy==1.26.4
packaging==24.0
pandas==2.2.2
pathlib==1.0.1
Pillow
portend==3.1.0
prompt-toolkit==3.0.43
proto-plus==1.24.0
protobuf==5.28.2
prov==2.0.1
psycopg2-binary==2.9.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pycryptodome==3.15.0
pydantic==1.10.2
pydot==2.0.0
PyHamcrest==2.0.2
PyJWT==1.7.1
PyMuPDF==1.24.5
PyMuPDFb==1.24.3
PyMySQL==1.0.2
pyOpenSSL==22.1.0
pyparsing==3.1.2
PyPDF2==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pythonnet==3.0.0rc4
pytz==2024.2
pyxnat==1.6.2
PyYAML==6.0.2
rdflib==6.3.2
reportlab==4.2.0
requests==2.28.1
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
scipy==1.13.1
simplejson==3.19.2
six==1.15.0
soupsieve==2.3.2.post1
sqlparse==0.3.1
starkbank-ecdsa==2.0.3
tempora==5.0.2
toolz==0.12.0
traits==6.3.2
txaio==20.4.1
typing_extensions==4.3.0
tzdata==2024.2
tzlocal==2.1
ublcii==1.0.3
uritemplate==4.1.1
urllib3==1.25.10
vine==5.1.0
wcwidth==0.2.13
XlsxWriter==3.0.3
zc.lockfile==2.0
zipp==3.2.0
zope.interface==5.2.0
gunicorn
urllib3>=1.25.10,<1.27
celery==5.3.4
redis==5.0.0
